feat: Update Activate and Transfer behaviors

+++ apps/models/models.py
@@ -80,7 +80,7 @@ class Material(models.Model):
     description   = models.TextField(_("Descripción"), blank=True)
     unit          = models.CharField(_("Unidad"), max_length=2, choices=categories.MATERIAL_UNITS, default="Unidades", blank=True,
                     help_text=_("Especifica la unidad a usar cuando se cuantifica este material."))
-    weight        = models.PositiveIntegerField(_("Peso unitario"), blank=True, null=True,
+    weight        = models.FloatField(_("Peso unitario"), blank=True, null=True,
                     help_text=_("Especifica el peso por unidad en kilogramos de manera aproximada. Se usará para hacer cálculos de materiales recuperados y puestos en uso"))
